Add files via upload
These are the scripts used to analyze the data for the publication doi: https://doi.org/10.1101/2022.04.26.489546. The scripts use the base function of R or R packages. The code are for grain width and can be implemented for other traits. Interested parties are welcome to contact if you require help to run the script.